implement show list of activitypackages
