+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="title_activitypackage_list">ActivityPackages</string>
     <string name="title_activitypackage_detail">ActivityPackage Detail</string>
+    <string name="task_done">Done</string>
 </resources>
